requirements.txt
streamlit
mediapipe
opencv-python
numpy
openai
pandas
python-dotenv
